move macro to utils module
