+++ gprimitives/src/lib.rs
@@ -39,17 +39,6 @@ use {
     },
 };
 
-/// Message handle.
-///
-/// Gear allows users and programs to interact with other users and programs via
-/// messages. Message creation consists of the following parts: message
-/// initialization, filling the message with payload (can be gradual), and
-/// message sending.
-#[repr(transparent)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, From, Into)]
-#[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
-pub struct MessageHandle(u32);
-
 /// The error type returned when conversion fails.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Display)]
 pub enum ConversionError {
@@ -64,116 +53,16 @@ pub enum ConversionError {
     InvalidSs58Address,
 }
 
-macro_rules! declare_primitive {
-    (@new $ty:ty) => {
-        impl $ty {
-            #[doc = concat!("Creates a new `", stringify!($ty), "` from a 32-byte array.")]
-            pub const fn new(array: [u8; 32]) -> Self {
-                Self(array)
-            }
-        }
-    };
-    (@zero $ty:ty) => {
-        impl $ty {
-            #[doc = concat!("Creates a new zero `", stringify!($ty), "`.")]
-            pub const fn zero() -> Self {
-                Self([0; 32])
-            }
-
-            #[doc = concat!("Checks whether `", stringify!($ty), "` is zero.")]
-            pub fn is_zero(&self) -> bool {
-                self == &Self::zero()
-            }
-        }
-    };
-    (@into_bytes $ty:ty) => {
-        impl $ty {
-            #[doc = concat!("Returns `", stringify!($ty), "`as bytes array.")]
-            pub fn into_bytes(self) -> [u8; 32] {
-                self.0
-            }
-        }
-    };
-    (@from_u64 $ty:ty) => {
-        impl From<u64> for $ty {
-            fn from(value: u64) -> Self {
-                let mut id = Self::zero();
-                id.0[..8].copy_from_slice(&value.to_le_bytes()[..]);
-                id
-            }
-        }
-    };
-    (@from_h256 $ty:ty) => {
-        #[cfg(feature = "codec")]
-        impl From<H256> for $ty {
-            fn from(h256: H256) -> Self {
-                Self::new(h256.to_fixed_bytes())
-            }
-        }
-    };
-    (@try_from_slice $ty:ty) => {
-        impl TryFrom<&[u8]> for $ty {
-            type Error = ConversionError;
-
-            fn try_from(slice: &[u8]) -> Result<Self, Self::Error> {
-                if slice.len() != 32 {
-                    return Err(ConversionError::InvalidSliceLength);
-                }
-
-                let mut ret = Self([0; 32]);
-                ret.as_mut().copy_from_slice(slice);
-
-                Ok(ret)
-            }
-        }
-    };
-    (@display $ty:ty) => {
-        impl fmt::Display for $ty {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                const LEN: usize = 32;
-                const MEDIAN: usize = (LEN + 1) / 2;
-
-                let mut e1 = MEDIAN;
-                let mut s2 = MEDIAN;
-
-                if let Some(precision) = f.precision() {
-                    if precision < MEDIAN {
-                        e1 = precision;
-                        s2 = LEN - precision;
-                    }
-                }
-
-                let mut out1 = [0; MEDIAN * 2];
-                let mut out2 = [0; MEDIAN * 2];
-
-                let _ = hex::encode_to_slice(&self.0[..e1], &mut out1[..e1 * 2]);
-                let _ = hex::encode_to_slice(&self.0[s2..], &mut out2[..(LEN - s2) * 2]);
-
-                let p1 = unsafe { str::from_utf8_unchecked(&out1[..e1 * 2]) };
-                let p2 = unsafe { str::from_utf8_unchecked(&out2[..(LEN - s2) * 2]) };
-                let sep = e1.ne(&s2).then_some("..").unwrap_or_default();
-
-                if f.alternate() {
-                    write!(f, "{}(0x{p1}{sep}{p2})", stringify!($ty))
-                } else {
-                    write!(f, "0x{p1}{sep}{p2}")
-                }
-            }
-        }
-    };
-    (@debug $ty:ty) => {
-        impl fmt::Debug for $ty {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                fmt::Display::fmt(self, f)
-            }
-        }
-    };
-    ($($feature:ident)*, $ty:ty) => {
-        $(
-            declare_primitive!(@$feature $ty);
-        )*
-    };
-}
+/// Message handle.
+///
+/// Gear allows users and programs to interact with other users and programs via
+/// messages. Message creation consists of the following parts: message
+/// initialization, filling the message with payload (can be gradual), and
+/// message sending.
+#[repr(transparent)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, From, Into)]
+#[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
+pub struct MessageHandle(u32);
 
 /// Program identifier.
 pub type ProgramId = ActorId;
@@ -191,7 +80,7 @@ pub type ProgramId = ActorId;
 #[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
 pub struct ActorId([u8; 32]);
 
-declare_primitive!(new zero into_bytes from_u64 from_h256 try_from_slice display debug, ActorId);
+utils::impl_primitive!(new zero into_bytes from_u64 from_h256 try_from_slice display debug, ActorId);
 
 impl ActorId {
     /// System program identifier.
@@ -246,7 +135,7 @@ impl FromStr for ActorId {
 #[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
 pub struct MessageId([u8; 32]);
 
-declare_primitive!(new zero into_bytes from_u64 from_h256 display debug, MessageId);
+utils::impl_primitive!(new zero into_bytes from_u64 from_h256 display debug, MessageId);
 
 impl MessageId {
     /// Generates `MessageId` for non-program outgoing message.
@@ -301,7 +190,7 @@ impl MessageId {
 #[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
 pub struct CodeId([u8; 32]);
 
-declare_primitive!(new zero into_bytes from_u64 from_h256 try_from_slice display debug, CodeId);
+utils::impl_primitive!(new zero into_bytes from_u64 from_h256 try_from_slice display debug, CodeId);
 
 impl CodeId {
     /// Generates `CodeId` from given code.
@@ -336,7 +225,7 @@ impl FromStr for CodeId {
 #[cfg_attr(feature = "codec", derive(TypeInfo, Encode, Decode, MaxEncodedLen), codec(crate = scale))]
 pub struct ReservationId([u8; 32]);
 
-declare_primitive!(new zero into_bytes from_u64 from_h256 display debug, ReservationId);
+utils::impl_primitive!(new zero into_bytes from_u64 from_h256 display debug, ReservationId);
 
 impl ReservationId {
     /// Generates `ReservationId` from given message and nonce.
